refactor: created registration DTOs
