+++ src/main/java/com/activecourses/upwork/dto/authentication/login/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.activecourses.upwork.dto;
+package com.activecourses.upwork.dto.authentication.login;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
