+++ src/main/java/com/activecourses/upwork/dto/authentication/registration/RegistrationRequestDto.java
@@ -1,0 +1,32 @@
+package com.activecourses.upwork.dto.authentication.registration;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class RegistrationRequestDto {
+    @NotEmpty(message = "First name is required")
+    @NotNull(message = "First name is required")
+    private String firstName;
+
+    @NotEmpty(message = "Last name is required")
+    @NotNull(message = "Last name is required")
+    private String lastName;
+
+    @Email(message = "Email is not valid")
+    @NotEmpty(message = "Email is required")
+    @NotNull(message = "Email is required")
+    private String email;
+
+    @NotEmpty(message = "Password is required")
+    @NotNull(message = "Password is required")
+    @Size(min = 6, message = "Password must be at least 6 characters long")
+    private String password;
+}
