+++ src/main/java/com/activecourses/upwork/dto/authentication/registration/RegistrationResponseDto.java
@@ -1,0 +1,12 @@
+package com.activecourses.upwork.dto.authentication.registration;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class RegistrationResponseDto {
+    private String message;
+}
